Checkpoint 2
Now seems to be working. Further work to do on Tile object - and adding create/delete/modify methods to restrictions ...
