+++ InstantPrint/InstantPrintTool.py
@@ -39,8 +39,6 @@ from ..constants import (
     RestrictionAction
 )
 
-from ..core.TOMsProposal import (TOMsProposal)
-
 class InstantPrintDialog(QDialog):
 
     hidden = pyqtSignal()
@@ -106,18 +104,6 @@ class InstantPrintTool(QgsMapTool, InstantPrintDialog):
                 if scale:
                     self.retrieve_scales(scale)
         self.check_scales()
-
-        # TH (180608) change signals to new functions
-        self.dialogui.comboBox_layouts.currentIndexChanged.disconnect(self.__selectLayout)
-        self.dialogui.comboBox_layouts.currentIndexChanged.connect(self.TOMsSelectLayout)
-
-        self.exportButton.clicked.disconnect(self.__export)
-        self.exportButton.clicked.connect(self.TOMsExport)
-
-        # self.iface.composerAdded.disconnect()
-        # self.iface.composerWillBeRemoved.disconnect(self.__reloadLayouts)
-        # self.iface.composerAdded.connect(lambda view: self.TOMsReloadLayouts())
-        # self.iface.composerWillBeRemoved.connect(self.TOMsReloadLayouts)
 
     def __onDialogHidden(self):
         self.setEnabled(False)
@@ -398,444 +384,3 @@ class InstantPrintTool(QgsMapTool, InstantPrintDialog):
                 self.dialogui.addScale.setEnabled(True)
                 self.dialogui.deleteScale.setVisible(False)
 
-    def TOMsSelectLayout(self):
-
-        # TH (180608): Need to be able to accept composers with multiple maps - so over ride original
-
-        if not self.dialog.isVisible():
-            return
-        activeIndex = self.dialogui.comboBox_layouts.currentIndex()
-        if activeIndex < 0:
-            return
-
-        layout_name = self.dialogui.comboBox_layouts.currentText()
-        self.layoutView = self.projectLayoutManager.layoutByName(layout_name)
-
-        layoutView = self.dialogui.comboBox_layouts.itemData(activeIndex)
-
-        self.exportButton.setEnabled(True)
-
-        #self.layoutView = layoutView
-        #self.mapitem = maps[0]
-
-        if self.layoutView.atlas():
-            #if len(maps) != 1:
-            #QMessageBox.warning(self.iface.mainWindow(), self.tr("Invalid composer"), self.tr("The composer must have exactly one map item."))
-            #self.exportButton.setEnabled(False)
-            self.iface.mapCanvas().scene().removeItem(self.rubberband)
-            self.rubberband = None
-            self.dialogui.comboBox_scale.setEnabled(False)
-        else:
-            self.dialogui.comboBox_scale.setEnabled(True)
-            #self.dialogui.comboBox_scale.setScale(1 / self.mapitem.scale())
-            self.dialogui.comboBox_scale.setScale(1 / 500)  # composer may not have a "scale" item
-            self.__createRubberBand()
-
-    def TOMsExport(self):
-
-        # TH (180608): Export function to deal with atlases
-
-        settings = QSettings()
-        format = self.dialogui.comboBox_fileformat.itemData(self.dialogui.comboBox_fileformat.currentIndex())
-
-        # TH (180608): Check to see whether or not the Composer is an Atlas
-        currPrintLayout = self.layoutView
-        # currLayoutAtlas = currPrintLayout.atlasComposition()
-
-        self.proposalForPrintingStatusText = "PROPOSED"
-        self.proposalPrintTypeDetails = "Print Date"
-        self.openDateForPrintProposal = self.proposalsManager.date()
-
-        # TODO: Move to plugin wide variable for datasource
-        self.Proposals = self.tableNames.setLayer("Proposals")
-        """if QgsProject.instance().mapLayersByName("Proposals"):
-            self.Proposals = \
-                QgsProject.instance().mapLayersByName("Proposals")[0]
-        else:
-            QMessageBox.information(self.iface.mainWindow(), "ERROR",
-                                    ("Table Proposals is not present"))"""
-
-        if currPrintLayout.atlas():
-
-            if self.proposalsManager.currentProposal() == 0:
-
-                # Include a dialog to check which proposal is required - or everything
-
-                self.proposalForPrintingStatusText = "CONFIRMED"
-                self.proposalPrintTypeDetails = "Effective Date"
-
-                # set the dialog (somehow)
-                self.acceptedProposalDialog = acceptedProposalsDialog()
-
-                self.createAcceptedProposalcb()
-                self.acceptedProposalDialog.show()
-
-                # Run the dialog event loop
-                result = self.acceptedProposalDialog.exec_()
-                # See if OK was pressed
-                if result:
-
-                    # Take the output from the form and set the current Proposal
-                    indexProposal = self.acceptedProposalDialog.cb_AcceptedProposalsList.currentIndex()
-                    proposalNrForPrinting = self.acceptedProposalDialog.cb_AcceptedProposalsList.itemData(indexProposal)
-                    printProposal = TOMsProposal(self.proposalsManager, proposalNrForPrinting)
-                    # self.openDateForPrintProposal = self.proposalsManager.getProposalOpenDate(proposalNrForPrinting)
-                    self.openDateForPrintProposal = printProposal.getProposalOpenDate()
-                else:
-                    return
-
-            else:
-
-                proposalNrForPrinting = self.proposalsManager.currentProposal()
-
-            self.TOMsExportAtlas(proposalNrForPrinting)
-
-        else:
-
-            self.__export()
-
-
-    def TOMsExportAtlas(self, currProposalID):
-
-        # TH (180608): Export function to deal with atlases
-
-        settings = QSettings()
-        format = self.dialogui.comboBox_fileformat.itemData(self.dialogui.comboBox_fileformat.currentIndex())
-
-        # TH (180608): Check to see whether or not the Composer is an Atlas
-        currPrintLayout = self.layoutView
-        currLayoutAtlas = currPrintLayout.atlas()
-
-        success = False
-
-        # https://gis.stackexchange.com/questions/77848/programmatically-load-composer-from-template-and-generate-atlas-using-pyqgis?utm_medium=organic&utm_source=google_rich_qa&utm_campaign=google_rich_qa
-
-        self.TOMsSetAtlasValues(currPrintLayout)
-        
-        currRevisionDate = self.proposalsManager.date()
-
-        # get the map tiles that are affected by the Proposal
-        # self.getProposalTileList(currProposalID, currRevisionDate)
-        proposalTileListForDate = TOMsProposal(self.proposalsManager, self.proposalsManager.currentProposal()).getProposalTileListForDate(currRevisionDate)
-        self.tileSet = set(proposalTileListForDate)  # TODO: Change around use of tileSet - also might be good to have a current proposal as an object in proposalManager...
-
-        # Now check which tiles to use
-        self.TOMsChooseTiles()
-
-        if len(self.tilesToPrint) == 0:
-            return
-
-        # get the output location
-        dirName = QFileDialog.getExistingDirectory(
-            self.iface.mainWindow(),
-            self.tr("Export Composition"),
-            settings.value("/instantprint/lastdir", ""),
-            QFileDialog.ShowDirsOnly
-        )
-        if not dirName:
-            return
-
-        settings.setValue("/instantprint/lastdir", dirName)
-
-        tileIDList = ""
-        firstTile = True
-        for tile in self.tilesToPrint:
-            if firstTile:
-                tileIDList = str(tile.attribute("id"))
-                firstTile = False
-            else:
-                tileIDList = tileIDList + ',' + str(tile.attribute("id"))
-
-        currLayoutAtlas.setFilterFeatures(True)
-        currLayoutAtlas.setFilterExpression(' "id" in ({tileList})'.format(tileList=tileIDList))
-
-        composerRevisionNr = currPrintLayout.itemById('revisionNr')
-        composerEffectiveDate = currPrintLayout.itemById('effectiveDate')
-        composerProposalStatus = currPrintLayout.itemById('proposalStatus')
-
-        if composerProposalStatus is not None:
-            composerProposalStatus.setText(self.proposalForPrintingStatusText)
-        else:
-            QMessageBox.warning(self.iface.mainWindow(), self.tr("Missing label in Layout"),
-                                self.tr("Missing label 'proposalStatus'"))
-
-        composerPrintTypeDetails = currPrintLayout.itemById('printTypeDetails')
-        if composerPrintTypeDetails is not None:
-            composerPrintTypeDetails.setText(self.proposalPrintTypeDetails)
-        else:
-            QMessageBox.warning(self.iface.mainWindow(), self.tr("Missing label in Layout"),
-                                self.tr("Missing label 'printTypeDetails'"))
-
-            # currProposalTitle, currProposalOpenDate = self.getProposalTitle(currProposalID)
-        printProposal = TOMsProposal(self.proposalsManager, currProposalID)
-        currProposalTitle = printProposal.getProposalTitle()
-        currProposalOpenDate = printProposal.getProposalOpenDate()
-
-        if currProposalID == 0:
-            currProposalTitle = "CurrentRestrictions_({date})".format(date=self.proposalsManager.date().toString('yyyyMMMdd'))
-
-        QgsMessageLog.logMessage("In TOMsExportAtlas. Now printing " + str(currLayoutAtlas.count()) + " items ....", tag="TOMs panel")
-
-        currLayoutAtlas.setEnabled(True)
-        currLayoutAtlas.updateFeatures()
-        currLayoutAtlas.beginRender()
-
-        altasFeatureFound = currLayoutAtlas.first()
-
-        while altasFeatureFound:
-
-            currTileNr = int(currLayoutAtlas.nameForPage(currLayoutAtlas.currentFeatureNumber()))
-
-            currLayoutAtlas.refreshCurrentFeature()
-
-            tileWithDetails = self.tileFromTileSet(currTileNr)
-
-            if tileWithDetails == None:
-                QgsMessageLog.logMessage("In TOMsExportAtlas. Tile with details not found ....", tag="TOMs panel")
-                QMessageBox.warning(self.iface.mainWindow(), self.tr("Print Failed"),
-                                    self.tr("Could not find details for " + str(currTileNr)))
-                break
-
-            QgsMessageLog.logMessage("In TOMsExportAtlas. tile nr: " + str(currTileNr) + " RevisionNr: " + str(tileWithDetails["RevisionNr"]) + " RevisionDate: " + str(tileWithDetails["LastRevisionDate"]), tag="TOMs panel")
-
-            if self.proposalForPrintingStatusText == "CONFIRMED":
-                composerRevisionNr.setText(str(tileWithDetails["RevisionNr"]))
-                composerEffectiveDate.setText(
-                    '{date}'.format(date=tileWithDetails["LastRevisionDate"].toString('dd-MMM-yyyy')))
-            else:
-                composerRevisionNr.setText(str(tileWithDetails["RevisionNr"]+1))
-                # For the Proposal, use the current view date
-                composerEffectiveDate.setText(
-                    '{date}'.format(date=self.openDateForPrintProposal.toString('dd-MMM-yyyy')))
-
-            filename = currProposalTitle + "_" + str(currTileNr) + "." + self.dialogui.comboBox_fileformat.currentText().lower()
-            outputFile = os.path.join(dirName, filename)
-
-            exporter = QgsLayoutExporter(currLayoutAtlas.layout())
-
-            if self.dialogui.comboBox_fileformat.currentText().lower() == u"pdf":
-                result = exporter.exportToPdf (outputFile, QgsLayoutExporter.PdfExportSettings())
-                # success = currLayoutAtlas.composition().exportAsPDF(outputFile)
-            else:
-                result = exporter.exportToImage(outputFile, 'png', QgsLayoutExporter.ImageExportSettings())
-                """image = currLayoutAtlas.composition().printPageAsRaster(0)
-                if not image.isNull():
-                    success = image.save(outputFile)"""
-
-            if result != QgsLayoutExporter.Success:
-                QMessageBox.warning(self.iface.mainWindow(), self.tr("Print Failed"),
-                                    self.tr("Failed to print " + exporter.errorFile()))
-                break
-
-            altasFeatureFound = currLayoutAtlas.next()
-
-        currLayoutAtlas.endRender()
-
-        QMessageBox.information(self.iface.mainWindow(), "Information",
-                                ("Printing completed"))
-
-    def tileFromTileSet(self, tileNr):
-
-        for tile in self.tileSet:
-            if tile.attribute("id") == tileNr:
-                return tile
-
-        return None
-
-    def TOMsReloadLayouts(self, removed=None):
-        if not self.dialog.isVisible():
-            # Make it less likely to hit the issue outlined in https://github.com/qgis/QGIS/pull/1938
-            return
-
-        self.dialogui.comboBox_layouts.blockSignals(True)
-        prev = None
-        if self.dialogui.comboBox_layouts.currentIndex() >= 0:
-            prev = self.dialogui.comboBox_layouts.currentText()
-        self.dialogui.comboBox_layouts.clear()
-        active = 0
-        for composer in self.iface.activeComposers():
-            if composer != removed and composer.composerWindow():
-                cur = composer.composerWindow().windowTitle()
-                self.dialogui.comboBox_layouts.addItem(cur, composer)
-                if prev == cur:
-                    active = self.dialogui.comboBox_layouts.count() - 1
-        self.dialogui.comboBox_layouts.setCurrentIndex(-1)  # Ensure setCurrentIndex below actually changes an index
-        self.dialogui.comboBox_layouts.blockSignals(False)
-        if self.dialogui.comboBox_layouts.count() > 0:
-            self.dialogui.comboBox_layouts.setCurrentIndex(active)
-
-            # TH (180608): Check whether or not the active item is an Atlas
-            layoutView = self.dialogui.comboBox_layouts.itemData(active)
-            currPrintLayout = layoutView.composition()
-            currLayoutAtlas = currPrintLayout.atlasComposition()
-
-            if currLayoutAtlas.enabled():
-                self.dialogui.comboBox_scale.setEnabled(False)
-            else:
-                self.dialogui.comboBox_scale.setEnabled(True)
-
-            self.exportButton.setEnabled(True)
-        else:
-            self.exportButton.setEnabled(False)
-            self.dialogui.comboBox_scale.setEnabled(False)
-
-    def TOMsSetAtlasValues(self, currPrintLayout):
-
-        # Need to set date and status
-
-        composerEffectiveDate = currPrintLayout.itemById('effectiveDate')
-
-        composerProposalStatus = currPrintLayout.itemById('proposalStatus')
-        composerProposalStatus.setText(self.proposalForPrintingStatusText)
-
-    def TOMsChooseTiles(self):
-        # function to display and allow choice of tiles for printing
-
-        QgsMessageLog.logMessage("In TOMsChooseTiles ", tag="TOMs panel")
-
-        # set the dialog (somehow)
-
-        self.tilesToPrint = []
-        idxMapTileId = self.tableNames.setLayer("MapGrid").fields().indexFromName("id")
-
-        self.tileListDialog = printListDialog(self.tileSet, idxMapTileId)
-
-        self.tileListDialog.show()
-
-        # Run the dialog event loop
-        result = self.tileListDialog.exec_()
-        # See if OK was pressed
-        if result:
-            QgsMessageLog.logMessage("In TOMsChooseTiles. Now printing - getValues ...",
-                                     tag="TOMs panel")
-            self.tilesToPrint = self.tileListDialog.getValues()
-
-        # ToDo: deal with cancel
-
-        # https://stackoverflow.com/questions/46057737/dynamically-changeable-qcheckbox-list
-        pass
-
-class printListDialog(printListDialog, QDialog):
-    def __init__(self, initValues, idxValue, parent=None):
-        QDialog.__init__(self, parent)
-
-        # initValues is set of features (in this case MapTile features); idxValue is the index to the set
-        # Set up the user interface from Designer.
-        self.setupUi(self)
-
-        QgsMessageLog.logMessage("In printListDialog. Initiating ...",             tag="TOMs panel")
-
-        self.initValues = initValues
-        self.idxValue = idxValue
-        self.tilesToPrint = self.initValues.copy()
-
-        self.LIST = self.findChild(QListWidget, "tileList")
-        self.cbToggleTiles = self.findChild(QCheckBox, "cb_ToggleTiles")
-        self.buttonBox = self.findChild(QDialogButtonBox, "buttonBox")
-        self.cbToggleTiles.setCheckState(Qt.Checked)
-
-        QMetaObject.connectSlotsByName(self)
-
-        self.populateTileListDialog()
-
-        self.LIST.itemChanged.connect(self.changeValues)
-        self.cbToggleTiles.stateChanged.connect(self.toggleValues)
-
-    def changeValues(self, element):
-        '''Whenever a checkbox is checked, modify the values'''
-        # Check if we check or uncheck the value:
-        """QgsMessageLog.logMessage(
-            "In printListDialog. In changeValues for: " + str(element.data(Qt.DisplayRole)),
-            tag="TOMs panel")"""
-
-        if element.checkState() == Qt.Checked:
-            #self.tilesToPrint.append(element.data(Qt.DisplayRole))
-            self.addFeatureToSet(element.data(Qt.DisplayRole))
-        else:
-            self.removedFeatureFromSet(element.data(Qt.DisplayRole))
-
-    def addFeatureToSet(self, valueToAdd):
-
-        for feature in sorted(self.initValues, key=lambda f: f[self.idxValue]):
-            attr = feature.attributes()
-            currValue = attr[self.idxValue]
-
-            if int(currValue) == int(valueToAdd):
-                self.tilesToPrint.add(feature)
-                QgsMessageLog.logMessage(
-                    "In TOMsTileListDialog. Adding: " + str(currValue) + " ; " + str(len(self.tilesToPrint)),
-                    tag="TOMs panel")
-                return
-
-    def removedFeatureFromSet(self, valueToRemove):
-
-        for feature in sorted(self.initValues, key=lambda f: f[self.idxValue]):
-            attr = feature.attributes()
-            currValue = attr[self.idxValue]
-
-            if int(currValue) == int(valueToRemove):
-                self.tilesToPrint.remove(feature)
-                QgsMessageLog.logMessage(
-                    "In TOMsTileListDialog. Removing: " + str(currValue) + " ; " + str(len(self.tilesToPrint)) ,
-                    tag="TOMs panel")
-                return
-
-    def toggleValues(self):
-        '''Whenever a checkbox is checked, modify the values'''
-        # Check if we check or uncheck the value:
-
-        QgsMessageLog.logMessage(
-            "In TOMsTileListDialog. In toggleValues. toggleState: " + str(self.cbToggleTiles.checkState()),
-            tag="TOMs panel")
-
-        if self.cbToggleTiles.checkState() == Qt.Checked:
-            for i in range(self.LIST.count()):
-                # attr = feature.attributes()
-                element = QListWidgetItem(self.LIST.item(i).text())
-                self.LIST.item(i).setCheckState(Qt.Checked)
-                """QgsMessageLog.logMessage(
-                    "In TOMsTileListDialog. In toggleValues. setting: " + str(self.LIST.item(i).text()),
-                    tag="TOMs panel")"""
-
-        else:
-
-            for i in range(self.LIST.count()):
-                # attr = feature.attributes()
-                element = QListWidgetItem(self.LIST.item(i).text())
-                self.LIST.item(i).setCheckState(Qt.Unchecked)
-
-    def populateTileListDialog(self, txtFilter=None):
-        '''Fill the QListWidget with values'''
-        # Delete everything
-        QgsMessageLog.logMessage("In TOMsTileListDialog. In populateList",
-                                 tag="TOMs panel")
-
-        self.LIST.clear()
-
-        for feature in sorted(self.initValues, key=lambda f: f[self.idxValue]):
-            attr = feature.attributes()
-            value = attr[self.idxValue]
-            element = QListWidgetItem(str(value))
-            element.setData(Qt.UserRole, attr[self.idxValue])
-
-            """QgsMessageLog.logMessage("In populateTileListDialog. Set State for " + str(attr[self.idxValue]),
-                                     tag="TOMs panel")"""
-
-            self.tilesToPrint.add(feature)
-            element.setCheckState(Qt.Checked)
-
-            self.LIST.addItem(element)
-
-    def getValues(self):
-        '''Return the selected values of the QListWidget'''
-
-        QgsMessageLog.logMessage("In TOMsTileListDialog. In getValues. Len List = " + str(len(self.tilesToPrint)),
-                                 tag="TOMs panel")
-
-        for feature in sorted(self.tilesToPrint, key=lambda f: f[self.idxValue]):
-            attr = feature.attributes()
-            currValue = attr[self.idxValue]
-            QgsMessageLog.logMessage("In Choose tiles form ... Returning " + str(attr[self.idxValue]),
-                                     tag="TOMs panel")
-
-        return self.tilesToPrint
